app/requirements.txt
Django==3.1.7
djangorestframework==3.12.2
psycopg2-binary==2.8.6